// resources/js/Pages/Resume/types/TechSkill.ts
export enum TechSkill {
	JavaScript = 'JavaScript',
	PHP = 'PHP',
	Java = 'Java',
	Python = 'Python',
	TypeScript = 'TypeScript',
	CSS = 'CSS',
	React = 'React',
	Laravel = 'Laravel',
	Bootstrap = 'Bootstrap',
	Jquery = 'jQuery',
	Tailwind = 'Tailwind',
	NextJS = 'NextJS',
	MySQL = 'MySQL',
	SASS = 'SASS',
	WordPress = 'WordPress',
	C = 'C',
	ShellScript = 'Shell Script',
	Bash = 'Bash',
	Docker = 'Docker',
	HTML = 'HTML',
	CSharp = 'C#',
	DotNet = 'DotNet',
	NodeJS = 'NodeJS',
	DynamoDB = 'DynamoDB',
	Blazor = 'Blazor',
	AWS = 'AWS',
	GitHub = 'GitHub',
	MicrosoftAzure = 'Microsoft Azure',
	PowerShell = 'PowerShell'
}